auth: refactor JwtTokenProvider → JwtTokenService with interface + impl
