+++ src/main/java/com/challengeteam/shop/security/filter/JwtTokenFilter.java
@@ -1,6 +1,6 @@
 package com.challengeteam.shop.security.filter;
 
-import com.challengeteam.shop.security.JwtTokenProvider;
+import com.challengeteam.shop.service.JwtTokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -17,7 +17,7 @@ import java.io.IOException;
 @RequiredArgsConstructor
 public class JwtTokenFilter extends OncePerRequestFilter {
 
-    private final JwtTokenProvider jwtTokenProvider;
+    private final JwtTokenService jwtTokenService;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
@@ -25,8 +25,8 @@ public class JwtTokenFilter extends OncePerRequestFilter {
         if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
             bearerToken = bearerToken.substring(7);
         }
-        if (bearerToken != null && jwtTokenProvider.isValid(bearerToken)) {
-            Authentication authentication = jwtTokenProvider.getAuthentication(bearerToken);
+        if (bearerToken != null && jwtTokenService.isValid(bearerToken)) {
+            Authentication authentication = jwtTokenService.getAuthentication(bearerToken);
             if (authentication != null) {
                 SecurityContextHolder.getContext().setAuthentication(authentication);
             }
